+++ include/UserDataAPI.hpp
@@ -25,21 +25,30 @@
 #endif
 
 namespace user_data {
+    constexpr std::string_view ID = "hiimjasmine00.user_data_api";
+
     /// Returns whether the given node contains user data for the specified ID.
     /// @param node The node to check
     /// @param id The ID to check for (default is your mod ID)
     /// @returns True if the node contains user data for the specified ID, false otherwise.
     inline bool contains(cocos2d::CCNode* node, std::string_view id = GEODE_MOD_ID) {
-        return node->getUserObject(fmt::format("hiimjasmine00.user_data_api/{}", id)) != nullptr;
+        return node->getUserObject(fmt::format("{}/{}", ID, id)) != nullptr;
     }
 
-    /// Retrieves user data from the given node for the specified ID and converts it to the specified type.
+    /// Returns whether the given node is currently downloading user data.
+    /// @param node The node to check
+    /// @returns True if the node is currently downloading user data, false otherwise.
+    inline bool downloading(cocos2d::CCNode* node) {
+        return node->getUserObject(fmt::format("{}/downloading", ID)) != nullptr;
+    }
+
+    /// Retrieves user data from the given node for the specified ID and converts it to the specified type. (Defaults to matjson::Value)
     /// @param node The node to retrieve data from
     /// @param id The ID to retrieve data for (Defaults to your mod ID)
     /// @returns A Result containing the user data as the specified type, or an error if the data could not be retrieved or converted.
     template <class T = matjson::Value>
     inline geode::Result<T> get(cocos2d::CCNode* node, std::string_view id = GEODE_MOD_ID) {
-        if (auto obj = static_cast<cocos2d::CCString*>(node->getUserObject(fmt::format("hiimjasmine00.user_data_api/{}", id)))) {
+        if (auto obj = static_cast<cocos2d::CCString*>(node->getUserObject(fmt::format("{}/{}", ID, id)))) {
             return matjson::parseAs<T>(obj->m_sString);
         }
         else {
@@ -55,6 +64,6 @@ namespace user_data {
 
 #ifdef USER_DATA_API_EVENTS
 inline void user_data::upload(const matjson::Value& data, std::string_view id) {
-    geode::DispatchEvent<matjson::Value, std::string>("hiimjasmine00.user_data_api/v1/upload", data, std::string(id)).post();
+    geode::DispatchEvent<matjson::Value, std::string>(fmt::format("{}/v1/upload", ID), data, std::string(id)).post();
 }
 #endif