+++ changelog.md
@@ -1,3 +1,7 @@
 # User Data API Changelog
-## v1.0.0 (2024-08-23)
+## v1.1.0 (2025-08-28)
+- Added `user_data::downloading` method to check if a node is currently downloading user data
+- Added `user_data::ID` constant for the mod ID
+
+## v1.0.0 (2025-08-23)
 - Initial release